requirements.txt
fastapi
uvicorn
langgraph
langchain-groq
beautifulsoup4
python-dotenv
httpx
gspread
oauth2client
